Notifications sending for alerts, invasions
notification progress


setting up notified_ids database


notifier work


notif work


Pinging!


Set up ping for type or item


fix for null notification


let them remove pings


turn off syndicates


fix track commands


fix darvo id not getting used


fix documentation


flub


some fixes, add pings to settings


some fixes


fixing some logger errors


fixes for logging and settings link


putt settings name back
Update Logger.js
rewards changes


remove ping on/off


fix delete after respond yes not working


logger fix


some changes for track stuff


fix for help


add start


added code blocks to start


command ids fetch fix


fallback for dm


fix for undefined guild id for dm's and forbidden error from reaction


more guild-specific fixes


fixes for bad promise


let someone know you can't set pings in direct message


pricecheck error fixes


clean up some errors


hopeful settings fix


split ping settings


small fix for no pings


don't send pings info if there are none


simplify logger


potential fix for making tracking easier
Update ShieldEmbed.js
fix errors around sending the shield embed
add kavat gene


Update Notifier.js
make news only those that are not updates or prime access
Update News.js
Update News.js
Update News.js
Update package.json
Send item picture along with item alerts


Update package.json
adding traces and kubrow eggs

+++ src/settings/Database.js
@@ -32,6 +32,7 @@ class Database {
     Object.assign(opts, dbOptions);
     this.db = mysql.createPool(opts);
     this.bot = bot;
+    this.logger = bot.logger;
 
     this.defaults = {
       prefix: '/',
@@ -63,7 +64,7 @@ class Database {
       .then((results) => {
         results.forEach((result) => {
           if (!result.isFulfilled()) {
-            this.bot.logger.error(result.reason());
+            this.logger.error(result.reason());
           }
         });
       });
@@ -261,9 +262,15 @@ class Database {
     .then((res) => {
       if (res[0].length === 0) {
         if (channel.type === 'text') {
-          return this.addGuildTextChannel(channel).then(() => this.defaults[`${setting}`]);
+          return this.addGuildTextChannel(channel)
+          .then(() => this.setChannelSetting(channel, setting, this.defaults[`${setting}`]))
+          .then(() => this.defaults[`${setting}`])
+          ;
         }
-        return this.addDMChannel(channel).then(() => this.defaults[`${setting}`]);
+        return this.addDMChannel(channel)
+        .then(() => this.setChannelSetting(channel, setting, this.defaults[`${setting}`]))
+        .then(() => this.defaults[`${setting}`])
+        ;
       }
       return res[0][0].val;
     });
@@ -380,14 +387,50 @@ class Database {
   }
 
   /**
+   * Get all pings for a guild for the provided event types and items
+   * @param  {Guild} guild                 Guild to get pings for
+   * @param  {Array.<string>} itemsOrTypes array of strings corresponding to event and reward types
+   * @returns {Promise.<string>}            Promise of a string to prepend to a message
+   */
+  getPing(guild, itemsOrTypes) {
+    const query = SQL`SELECT text FROM pings WHERE guild_id=${guild.id} AND item_or_type in (${itemsOrTypes})`;
+    return this.db.query(query)
+    .then((res) => {
+      if (res[0].length === 0) {
+        return '';
+      }
+      return res[0]
+        .map(result => result.text).join(', ');
+    });
+  }
+
+  getPingsForGuild(guild) {
+    if (guild) {
+      const query = SQL`SELECT item_or_type, text FROM pings WHERE guild_id=${guild.id}`;
+      return this.db.query(query)
+      .then((res) => {
+        if (res[0].length === 0) {
+          return '';
+        }
+        return res[0]
+          .map(result => ({ text: result.text, thing: result.item_or_type }));
+      });
+    }
+    return new Promise(resolve => resolve([]));
+  }
+
+  /**
    * Removes a ping message
    * @param {Guild} guild The guild where the ping message is currently being sent
    * @param {string} itemOrType The item or event type associated to the ping message
    * @returns {Promise}
    */
   removePing(guild, itemOrType) {
-    const query = SQL`DELETE FROM pings WHERE guild_id = ${guild.id} AND item_or_type = ${itemOrType};`;
-    return this.db.query(query);
+    if (guild) {
+      const query = SQL`DELETE FROM pings WHERE guild_id = ${guild.id} AND item_or_type = ${itemOrType};`;
+      return this.db.query(query);
+    }
+    return new Promise(resolve => resolve(false));
   }
 
   /**
@@ -395,25 +438,27 @@ class Database {
    * @param {string} type The type of the event
    * @param {string} platform The platform of the event
    * @param {Array.<string>} items The items in the reward that is being notified
-   * @returns {Promise.<Array.<{channel_id: string, webhook: string, ping: string}>>}
+   * @returns {Promise.<Array.<{channel_id: string, webhook: string}>>}
    */
   getNotifications(type, platform, items) {
-    const query = SQL`SELECT channel.id AS channel_id, channel.webhook AS webhook,
-      GROUP_CONCAT(pings.text SEPARATOR '\n') AS ping
-      FROM type_notifications ```
-      .append(items ? 'LEFT JOIN item_notifications USING (channel_id) ' : '')
-      .append('INNER JOIN channels ON type_notifications.channel_id = channels.id')
-      .append('LEFT JOIN pings ON channels.guild_id = pings.guild_id AND ( ')
-      .append(items ? '(item_notifications.item = pings.item_or_type AND item_notifications.ping = TRUE) OR ' : '')
-      .append(SQL```(type_notifications.type = pings.item_or_type AND type_notifications.ping = TRUE)
-      )
-      WHERE
-        type_notifications.type = ${type} AND
-        (IFNULL(channels.guild_id, 0) >> 22) % ${this.bot.shardCount} = ${this.bot.shardId} AND
-        channels.platform = ${platform} ```)
-      .append(items ? SQL```AND notifications.item IN (${items}) ``` : '')
-      .append('GROUP BY notifications.channel_id;');
-    return this.db.query(query);
+    try {
+      const query = SQL`SELECT DISTINCT channels.id as channelId
+          FROM type_notifications`
+      .append(items && items.length > 0 ?
+              SQL` INNER JOIN item_notifications ON type_notifications.channel_id = item_notifications.channel_id` : SQL``)
+      .append(SQL` INNER JOIN channels ON channels.id = type_notifications.channel_id`)
+      .append(SQL` INNER JOIN settings ON channels.id = settings.channel_id`)
+      .append(SQL`
+        WHERE type_notifications.type = ${String(type)}
+          AND MOD(IFNULL(channels.guild_id, 0) >> 22, ${this.bot.shardCount}) = ${this.bot.shardId}
+          AND settings.setting = "platform"  AND settings.val = ${platform || 'pc'} `)
+      .append(items && items.length > 0 ? SQL`AND item_notifications.item IN (${items})
+          AND item_notifications.channel_id = settings.channel_id;` : SQL`;`);
+      return this.db.query(query);
+    } catch (e) {
+      this.logger.error(e);
+      return [];
+    }
   }
 
   /**
@@ -629,6 +674,11 @@ class Database {
       });
   }
 
+  /**
+   * Remove guild from database
+   * @param  {snowflake} guild Guild to be removed from database
+   * @returns {Promise.<string>} status of removal
+   */
   removeGuild(guild) {
     const channelIds = guild.channels.keyArray();
     const permissionResults = [];
@@ -646,28 +696,81 @@ class Database {
     .then(res => res);
   }
 
+  /**
+   * Remove permissions corresponding to the guild id
+   * @param  {snowflake} guildId guild identifier for removal
+   * @returns {Promise.<string>} status of removal
+   */
   removeGuildPermissions(guildId) {
     const query = SQL`DELETE FROM guild_permissions WHERE guild_id = ${guildId}`;
     return this.db.query(query)
       .then(res => res);
   }
 
+  /**
+   * Remove permissions corresponding to the guild id
+   * @param  {snowflake} channelId channel identifier for removal
+   * @returns {Promise.<string>} status of removal
+   */
   removeChannelPermissions(channelId) {
     const query = SQL`DELETE FROM channel_permisions WHERE channel_id = ${channelId}`;
     return this.db.query(query)
     .then(res => res);
   }
 
+  /**
+   * Remove permissions corresponding to the channel id
+   * @param  {snowflake} channelId channel identifier for removal
+   * @returns {Promise.<string>} status of removal
+   */
   removeItemNotifications(channelId) {
     const query = SQL`DELETE FROM item_notifications WHERE channel_id = ${channelId}`;
     return this.db.query(query)
     .then(res => res);
   }
 
+  /**
+   * Remove pings corresponding to the guild id
+   * @param  {snowflake} guildId guild identifier for removal
+   * @returns {Promise.<string>} status of removal
+   */
   removePings(guildId) {
     const query = SQL`DELETE FROM pings WHERE guild_id = ${guildId}`;
     return this.db.query(query)
     .then(res => res);
+  }
+
+  /**
+   * Set the notified ids for a given platform and shard id
+   * @param {string} platform    platform corresponding to notified ids
+   * @param {number} shardId     shard id corresponding to notified ids
+   * @param {JSON} notifiedIds list of oids that have been notifiedIds
+   * @returns {Promise}
+   */
+  setNotifiedIds(platform, shardId, notifiedIds) {
+    const query = SQL`INSERT INTO notified_ids VALUES
+      (${shardId}, ${platform}, JSON_ARRAY(${notifiedIds}))
+      ON DUPLICATE KEY UPDATE id_list = JSON_ARRAY(${notifiedIds});`;
+    return this.db.query(query);
+  }
+
+  /**
+   * Get list of notified ids for the given platform and shard id
+   * @param  {string} platform Platform
+   * @param  {number} shardId  Identifier of the corresponding shard
+   * @returns {Promise.<Array>} Array of notified oids
+   */
+  getNotifiedIds(platform, shardId) {
+    const query = SQL`SELECT id_list
+      FROM notified_ids
+      WHERE shard_id=${shardId} AND platform=${platform};`;
+    return this.db.query(query)
+    .then((res) => {
+      if (res[0].length === 0) {
+        return [];
+      }
+      return res[0][0].id_list;
+    });
   }
 }
